List of functions in DG_Playground module

DG_Playground.Field1D — Type

Field1D(mesh)

Description

initialize Field1D struct

Arguments

mesh: a mesh to compute on

Return Values:

u : the field to be computed

u̇: numerical solutions for the field

flux: the numerical flux for the computation

DG_Playground.Mesh — Method

mesh(K, n, xmin, xmax)

Description

outer_constructor mesh struct

Arguments

K: number of elements

n: polynomial order

xmin: lower bound

xmax: upper bound

Return Values: x

return grid values

DG_Playground.external_params — Type

external_params{T,S}

Description

struct for external params needed for advection

Members

first is velocity
second is value for α

DG_Playground.build_operator — Method

buildoperator(affineoperator!, 𝒢::Mesh; mass_matrix = false)

Description

builds affine operator associated with DG operator

Comment

With non-homogenous boundary conditions becomes affine

DG_Playground.buildmaps1D — Method

buildmaps1D(K, nGL, nFP, nFaces, fmask, EtoE, EtoF, x)

Description

connectivity matrices for element to elements and elements to face

Arguments

K: number of elements

nGL: number of points within an element (polynomial degree + 1)

nFP: 1

nFaces: 2

fmask: an element by element mask to extract edge values

EtoE: element to element connectivity

EtoF: element to face connectivity

x: Guass lobatto points

Return Values: vmapM, vmapP, vmapB, mapB, mapI, mapO, vmapI, vmapO

vmapM: vertex indices, (used for interior u values)

vmapP: vertex indices, (used for exterior u values)

vmapB: vertex indices, corresponding to boundaries

mapB: use to extract vmapB from vmapM

mapI: Index of left boundary condition

mapO: Index of right boundary condition

Example | uses ../utils.jl

K = 3 n = 3; α = 0; β = 0; xmin = 0; xmax = 2π; nGL = n + 1 nFP = 1 nFaces = 2

r = jacobiGL(α, β, n)

VX, EtoV = unimesh1D(xmin, xmax, K) EtoE, EtoF = connect1D(EtoV) x = gridvalues1D(VX, EtoV, r) fx = edgevalues1D(r,x)

vmapM, vmapP, vmapB, mapB, mapI, mapO, vmapI, vmapO = buildmaps1D(K, nGL, nFP, nFaces, fmask, EtoE, EtoF, x)

DG_Playground.connect1D — Method

connect1D(EtoV)

Description

builds global connectivity arrays for 1D

Arguments

EtoV: element to node connectivity | a Matrix of size Kx2

Return Values: EtoE, EtoF

EtoE: element to element connectivity
EtoF: element to face connectivity

Example

DG_Playground.djacobi — Method

djacobi(x, α, β, n)

Description

Evaluates the derivative of the jacobi polynomial at the point x

Arguments

x: point at which you will evaluate the derivative of the jacobi polynomial

α: first parameter for Jacobi polynomials

β: second parameter for Jacobi polynomials

n : order

Return

y: the derivative of the of the Jacobi polynomial

DG_Playground.dmatrix — Method

dmatrix(x, α, β, N)

Description

Return the differentiation matrix of order N at the values x
Allocates too much memory

Arguments

x: points at which to evaluate the Jacobi polynomials

α: first parameter for Jacobi polynomials

β: second paramater for Jacobi polynomials

N: maximum order of Jacobi polynomial to include

Return Values

D: the differentiation matrix

Example

See LegendreTests.jl

DG_Playground.dropϵzeros! — Method

dropϵzeros!(sparseMatrix, drop_criteria)

Description

Drops machine zeros in sparse matrix

Arguments

A: a sparse matrix

drop_criteria: criteria for dropping entries

return

nothing

DG_Playground.dropϵzeros! — Method

dropϵzeros!(sparseMatrix)

Description

Drops machine zeros in sparse matrix

Arguments

!A: a sparse matrix

return

nothing

DG_Playground.dvandermonde — Method

dvandermonde(x, α, β, N)

Description

Return the gradient of the vandermonde matrix of order N at the values x
Allocates a little bit of memory

Arguments

x: points at which to evaluate the Jacobi polynomials

α: first parameter for Jacobi polynomials

β: second paramater for Jacobi polynomials

N: maximum order of Jacobi polynomial to include

Return Values

vr: gradient of vandermonde matrix

Example

See LegendreTests.jl

DG_Playground.edgevalues1D — Method

edgevalues1D(fmask, x)

Description

calculates edge values

Arguments

fmask: face mask for GL edges

x:  physical coordinates of solution on each element

Return Values: x

fx: face values of x

Example | ../utils.jl

r = jacobiGL(0, 0, 4) x = gridvalues1D(VX, EtoV, r) fmask = fmask1D(r)[1] fx = edgevalues1D(fmask,x)

the locations of the edges in element 1 is fx[:, 1]

DG_Playground.fmask1D — Method

facemask1D(r)

Description

creates face mask

Arguments

r: GL points

Return Values: x

fmask1: standard facemask
fmask2: alternate form

Example | ../utils.jl

r = jacobiGL(0, 0, 4) fmask = fmask1D(r)

DG_Playground.geometric_factors — Method

geometric_factors(x, Dʳ)

Description

computes the geometric factors for local mappings of 1D elements

Arguments

x: physical coordinates of solution for each element

Dʳ:

Return Values: rx, J

rx: inverse jacobian

J: jacobian (in 1D a scalar)

Example

DG_Playground.gridvalues1D — Method

gridvalues1D(xmin, xmax, K)

Description

Generates physical gridpoints with each element

Arguments

VX: vertex values | an Array of size K+1

EtoV: element to node connectivity | a Matrix of size Kx2

r: LGL nodes in reference element | an array

Return Values: x

x: physical coordinates of solution

Example (uses ../utils.jl as well)

xmin = 0 xmax = 2π K = 4

call functions

VX, EtoV = unimesh1D(xmin, xmax, K) r = jacobiGL(0, 0, 4) x = gridvalues1D(VX, EtoV, r)

x[:,1] is the physical coordinates within the first element

for plotting

f(x) = sin(x) plot(x, f.(x))

scatter(x,f.(x)) tends to work better

DG_Playground.jacobi — Method

jacobi(x, α, β, n)

Description

Evaluates the jacobi polynomial at the point x

Arguments

x: point at which you will evaluate the jacobi polynomial

α: first parameter for Jacobi polynomials

β: second parameter for Jacobi polynomials

n : order

Return

y: the value of the of the Jacobi polynomial

DG_Playground.jacobiGL — Method

jacobiGL(α, β, N)

Description

Guass Labatto quadrature points for the Jacobi Polynomial (α,β)
The quadrature weights are computed as well (but not returned)

Arguments

α, β: Jacobi polynomial descriptors

N: order of quadrature

Return: x

x: quadrature points | array of size N+1

Examples

julia> x = jacobiGL(0, 0, 4)
5-element Array{Float64,1}:
 -1.0
 -0.6546536707079759
  4.440892098500626e-16
  0.6546536707079771
  1.0

DG_Playground.jacobiGQ — Method

jacobiGQ(α, β, N)

Description

Guass Quadrature points and weights for the Jacobi Polynomial (α,β)

Input

α, β: Jacobi polynomial descriptors N: order of quadrature points

Return: x,w

x: quadrature points | array of size N+1 w: quadrature weights | array of size N+1 #Example α = 0 β = 0 N = 4 x, w = jacobiGQ(α, β, N)

DG_Playground.lift1D — Method

lift1D(V, y) for computing fluxes helps compute a surface integral of a quantity note that the parentheses are necessary to prevent too much multiplcation the E function takes the surface integrals are presents it with respect to the full space inside an element the entire operator represents how fluxes flow into the interior of an element

DG_Playground.lift1D_v2 — Method

lift1D_v2(V, y) for computing fluxes nodal form helps compute a surface integral of a quantity note that the parantheses are necessary to prevent too much multiplcation the E function takes the surface integrals are presents it with respect to the full space inside an element the entire operator represents how fluxes flow into the interior of an element

DG_Playground.make_periodic1D! — Method

make_periodic1D!(vmapP, u)

Description

makes the grid periodic by modifying vmapP.
Assumes that the first node is connected to the last.

Arguments

vmapP: exterior vertex map
u: vertex vector

Return Values: none

Example

DG_Playground.normals1D — Method

normals1D(K)

Description

calculates face normals

Arguments

K: number of elements

Return Values: normals

normals: face normals along each grid

Example

DG_Playground.unimesh1D — Method

unimesh1D(xmin, xmax, K)

Description

Generates a uniform 1D mesh

Arguments

xmin: smallest value of array
xmax: largest values of array
K: number of elements in an array

Return Values: VX, EtoV

VX: vertex values | an Array of size K+1
EtoV: element to node connectivity | a Matrix of size Kx2

Example

xmin = -1 xmax = 1 K = 4 VX, EtoV = unimesh1D(xmin, xmax, K)

DG_Playground.vandermonde — Method

vandermonde(x, α, β, N)

Description

Return vandermonde matrix of order N at the values x
Allocates a little bit of memory

Arguments

x: points at which to evaluate the Jacobi polynomials

α: first parameter for Jacobi polynomials

β: second parameter for Jacobi polynomials

N: maximum order of Jacobi polynomial to include

Return Values

v: vandermonde matrix

Example

See LegendreTests.jl